requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
matplotlib>=3.7.0
Pillow>=9.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0
aiofiles>=23.2.1
albumentations>=1.4.0
opencv-python-headless>=4.8.0.74
segmentation-models-pytorch>=0.3.3
timm>=0.9.7


